+++ bin/.gitignore
@@ -7,3 +7,5 @@
 /saul.class
 /someTest.class
 /test3D.class
+/clipTest.class
+/demo.class
