+++ src/DwarfEngine/MathTypes/Vector3.java
@@ -69,7 +69,16 @@ public final class Vector3{
 		this.x = x;
 		this.y = y;
 		this.z = z;
-		
+	}
+	public Vector3(Vector3 other) {
+		this.x = other.x;
+		this.y = other.y;
+		this.z = other.z;
+		this.w = other.w;
+	}
+	public Vector3(Vector2 other) {
+		this.x = other.x;
+		this.y = other.y;
 	}
 	
 	public void addTo(Vector3 vec) {
