--- src/Renderer3D/TriangleRenderer/package-info.java
@@ -1,1 +1,0 @@
-package Renderer3D.TriangleRenderer;
